Refactor response parsing into a `FeedImageMapper`
