+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -22,24 +22,14 @@ public final class RemoteFeedLoader: FeedLoader {
 		client.get(from: url) { result in
 			switch result {
 			case let .success((data, response)):
-				if response.statusCode == 200, let _ = try? JSONDecoder().decode(Root.self, from: data) {
-					completion(.success([]))
-				} else {
+				do {
+					let images = try FeedImageMapper.createFeedImages(from: data, and: response)
+					completion(.success(images))
+				} catch {
 					completion(.failure(Error.invalidData))
 				}
 			case .failure: completion(.failure(Error.connectivity))
 			}
 		}
 	}
-}
-
-private struct RemoteFeedImage: Decodable {
-	let imageId: String
-	let imageDesc: String?
-	let imageLoc: String?
-	let imageUrl: URL
-}
-
-private struct Root: Decodable {
-	let items: [RemoteFeedImage]
 }
